+++ scripts/update-swagger-ui.sh
@@ -6,7 +6,7 @@ set -eu -o pipefail
 
 version="${1:-""}"
 if [ -z "$version" ]; then
-	echo "Missing 'version' argument from command, run as $0 <version>" >&2 && exit 1
+    echo "Missing 'version' argument from command, run as $0 <version>" >&2 && exit 1
 fi
 zip_name="v$version.zip"
 
@@ -16,6 +16,7 @@ echo "Update vendored Swagger UI"
 mv "$zip_name" ./utoipa-swagger-ui-vendored/res/
 sed -i "s|version: \`.*\`|version: \`$version\`|" ./utoipa-swagger-ui-vendored/README.md
 sed -i "s|version: \`.*\`|version: \`$version\`|" ./utoipa-swagger-ui-vendored/src/lib.rs
+sed -i "s|res/v.*\.zip|res/v$version.zip|" ./utoipa-swagger-ui-vendored/src/lib.rs
 
 echo "Update utoipa-swagger-ui Swagger UI version"
 sed -i "s|tags/v.*>|tags/v$version.zip>|" ./utoipa-swagger-ui/README.md