+++ utoipa-swagger-ui-vendored/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+* Update Swagger UI with latest release 5.7.14 (https://github.com/juhaku/utoipa/pull/1142)
 * Remove commit commit id from changelogs (https://github.com/juhaku/utoipa/pull/1077)
 
 ## 0.1.1 - May 22 2024